Zahlenbereich angepasst, int war ein bisschen zu klein

+++ Serie 5/atoi.h
@@ -1,3 +1,3 @@
 # define isValid(c) (c >= '0'&& c <= '9') || (c >= 'A' && c <= 'F') || (c >= 'a' && c <= 'f')
 
-int atoi (const char* s);
+long int atoi (const char* s);
